pyproject.toml
[project]
name = "mcp-hello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.13.1",
    "httpx>=0.28.1",
    "keyring>=25.7.0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
]
